refactor: 리포지토리 수정
