+++ module-infra/persistence-db/src/main/kotlin/demo/kotlinpractice/member/repository/MemberJPARepository.kt
@@ -1,7 +1,6 @@
 package demo.kotlinpractice.member.repository
 
-import demo.kotlinpractice.domain.member.domain.Member
-import demo.kotlinpractice.domain.member.infra.MemberEntity
+import demo.kotlinpractice.member.entity.MemberEntity
 import org.springframework.data.jpa.repository.JpaRepository
 import java.util.Optional
 
@@ -9,5 +8,5 @@ interface MemberJPARepository : JpaRepository<MemberEntity, Long> {
 
     fun existsByName(name: String): Boolean
 
-    fun findByName(name: String): Optional<Member>
+    fun findByName(name: String): Optional<MemberEntity>
 }
