+++ module-infra/persistence-db/src/main/kotlin/demo/kotlinpractice/member/repository/MemberJPARepositoryAdaptor.kt
@@ -1,31 +1,32 @@
 package demo.kotlinpractice.member.repository
 
-import demo.kotlinpractice.domain.member.domain.Member
-import demo.kotlinpractice.domain.member.domain.repository.MemberRepository
-import demo.kotlinpractice.domain.member.infra.MemberEntity
+import demo.kotlinpractice.error.exception.MemberNotFoundException
+import demo.kotlinpractice.member.domain.Member
+import demo.kotlinpractice.member.entity.MemberEntity
+import demo.kotlinpractice.member.port.out.MemberPersistencePort
 import org.springframework.stereotype.Repository
 import kotlin.jvm.optionals.getOrNull
 
 @Repository
 class MemberJPARepositoryAdaptor(
     private val memberJPARepository: MemberJPARepository
-) : MemberRepository {
+) : MemberPersistencePort {
 
-    override fun findById(id: Long): Member {
-        return memberJPARepository.findById(id).getOrNull()?.toDomain()
-            ?: throw IllegalArgumentException()
-    }
+    override fun findById(id: Long): Member =
+        memberJPARepository.findById(id)
+            .getOrNull()?.toDomain()
+            ?: throw MemberNotFoundException()
 
-    override fun existsByName(name: String): Boolean {
-        return memberJPARepository.existsByName(name)
-    }
+    override fun existsByName(name: String): Boolean =
+        memberJPARepository.existsByName(name)
 
-    override fun save(member: Member): Member {
-        return memberJPARepository.save(MemberEntity.from(member)).toDomain()
-    }
+    override fun save(member: Member): Member =
+        memberJPARepository.save(
+            MemberEntity.from(member)
+        ).toDomain()
 
-    override fun findByName(name: String): Member {
-        return memberJPARepository.findByName(name).getOrNull()
-            ?: throw IllegalArgumentException()
-    }
+    override fun findByName(name: String): Member =
+        memberJPARepository.findByName(name)
+            .getOrNull()?.toDomain()
+            ?: throw MemberNotFoundException()
 }
